requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.10.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
astroid==3.3.11
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
audioread==3.1.0
babel==2.17.0
beautifulsoup4==4.13.5
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
Bottleneck==1.4.2
Brotli==1.0.9
bs4==0.0.2
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.1
cmudict==1.1.1
colorlog==6.9.0
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
coqpit-config==0.2.1
coqui-tts==0.27.2
coqui-tts-trainer==0.3.1
cycler==0.11.0
cymem==2.0.11
Cython==3.1.6
datasets==4.3.0
dateparser==1.1.8
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
Distance==0.1.3
distro==1.9.0
docopt==0.6.2
einops==0.8.1
emoji==1.2.0
en_core_web_sm==3.8.0
encodec==0.1.1
executing==2.2.1
Faker==37.8.0
fastjsonschema==2.21.2
fasttext-wheel==0.9.2
filelock==3.19.1
flake8==7.3.0
flatbuffers==25.9.23
fonts==0.0.3
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
gensim==4.4.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.75.1
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hangul-jamo==1.0.1
hangul-romanize==0.1.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
hyperopt==0.2.7
idna==3.10
imbalanced-learn==0.14.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==6.0.1
jamo==0.4.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jpype1==1.6.0
json5==0.12.1
jsonlines==1.2.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kobert-transformers==0.6.0
kollocate==0.0.2
konlpy==0.6.0
koparadigm==0.10.0
kss==6.0.5
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightgbm==4.6.0
llvmlite==0.45.1
lxml==6.0.2
Mako==1.3.10
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
monotonic-alignment-search==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.13
namex==0.1.0
narwhals==2.5.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
num2words==0.5.14
numba==0.62.1
numexpr==2.11.0
numpy==1.26.4
ollama==0.6.0
openai==1.108.1
openai-whisper==20250625
opt_einsum==3.4.0
optree==0.17.0
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pecab==1.0.8
pexpect==4.9.0
pillow==11.3.0
pip==25.2
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
pooch==1.8.2
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.28.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==18.1.0
pybind11==2.9.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.8
PyMySQL==1.1.2
pyparsing==3.2.3
pysbd==0.3.4
PySocks==1.7.1
pytest==8.4.2
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
srsly==2.5.1
stack-data==0.6.3
streamlit==1.51.0
streamlit-audiorecorder==0.0.6
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tossi==0.3.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
typeguard==4.4.4
typer==0.19.2
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
unicodedata2==15.1.0
Unidecode==1.4.0
uri-template==1.3.0
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
Whoosh==2.7.4
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.3
xgboost==3.0.5
xlrd==1.2.0
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2025.10.22
zipp==3.23.0
zstandard==0.23.0
